backend/requirements.txt
fastapi
bcrypt==4.0.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg[binary]
pydantic[email]
pydantic-settings
passlib[bcrypt]==1.7.4
alembic==1.12.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
minio==7.2.0
slowapi==0.1.9
python-dotenv==1.0.0